Ensure tags don't contain spaces when custom template is used
